@@ -2,8 +2,12 @@
     "private": true,
     "name": "wikidata-mismatch-finder",
     "scripts": {
-        "lint": "eslint --ext .js,.vue resources/js",
-        "lint:fix": "eslint --fix --ext .js,.vue resources/js",
+        "lint": "run-s lint:*",
+        "lint:eslint": "eslint --ext .js,.vue resources/js",
+        "lint:stylelint": "stylelint '**/*.scss' '**/*.vue'",
+        "fix": "run-s fix:*",
+        "fix:eslint": "eslint --fix --ext .js,.vue resources/js",
+        "fix:stylelint": "stylelint '**/*.scss' '**/*.vue' --fix",
         "i18n:validate": "banana-checker public/i18n",
         "test": "jest",
         "dev": "npm run development",
@@ -15,52 +19,51 @@
         "production": "mix --production"
     },
     "devDependencies": {
-        "@types/jest": "^29.5.5",
-        "@types/lodash": "^4.14.199",
-        "@typescript-eslint/eslint-plugin": "^6.7.5",
-        "@typescript-eslint/parser": "^6.7.5",
+        "@pinia/testing": "^0.1.3",
+        "@types/jest": "^29.5.12",
+        "@types/lodash": "^4.14.202",
+        "@typescript-eslint/eslint-plugin": "^6.20.0",
+        "@typescript-eslint/parser": "^6.20.0",
+        "@vue/compiler-sfc": "^3.4.15",
         "@vue/eslint-config-typescript": "^12.0.0",
-        "@vue/test-utils": "^1.3.6",
-        "@vue/vue2-jest": "^29.2.6",
-        "@wikimedia/codex": "^0.20.0",
-        "@wikimedia/codex-design-tokens": "^0.20.0",
-        "@wikimedia/codex-icons": "^0.20.0",
-        "@wikimedia/language-data": "^1.1.4",
-        "axios": "^1.5.1",
+        "@vue/runtime-dom": "^3.4.15",
+        "@vue/test-utils": "^2.4.4",
+        "@vue/vue3-jest": "^29.2.4",
+        "@wikimedia/codex": "^1.0.1",
+        "@wikimedia/codex-design-tokens": "^1.0.1",
+        "@wikimedia/codex-icons": "^1.0.1",
+        "@wikimedia/language-data": "^1.1.6",
+        "axios": "^1.6.7",
         "axios-mock-adapter": "^1.22.0",
         "babel-core": "^7.0.0-bridge.0",
-        "eslint": "^8.51.0",
-        "eslint-plugin-vue": "^9.17.0",
+        "eslint": "^8.56.0",
+        "eslint-plugin-vue": "^9.21.1",
         "grunt-banana-checker": "^0.11.1",
         "jest": "^29.7.0",
         "jest-environment-jsdom": "^29.7.0",
         "laravel-mix": "^6.0.49",
+        "npm-run-all": "^4.1.5",
         "postcss": "^8.4.31",
+        "postcss-html": "^1.6.0",
         "resolve-url-loader": "^5.0.0",
-        "sass": "^1.69.0",
-        "sass-loader": "^13.3.1",
-        "ts-jest": "^29.1.0",
+        "sass": "^1.70.0",
+        "sass-loader": "^14.1.0",
+        "stylelint": "^16.2.1",
+        "stylelint-config-recommended-vue": "^1.5.0",
+        "stylelint-config-standard-scss": "^13.0.0",
+        "ts-jest": "^29.1.2",
         "ts-loader": "^9.5.0",
-        "typescript": "^5.2.2",
-        "vue-loader": "^15.10.2",
-        "vue-template-compiler": "^2.6.14"
+        "typescript": "^5.3.3",
+        "vue-loader": "^17.3.1"
     },
     "dependencies": {
         "@inertiajs/inertia": "^0.11.0",
-        "@inertiajs/inertia-vue": "^0.8.0",
-        "@vue/composition-api": "^1.7.2",
-        "@wmde/wikit-tokens": "^2.1.0-alpha.15",
-        "@wmde/wikit-vue-components": "^2.1.0-alpha.16",
-        "date-fns": "^2.30.0",
+        "@inertiajs/inertia-vue3": "^0.6.0",
+        "date-fns": "^3.3.1",
         "lodash": "^4.17.21",
+        "pinia": "^2.1.7",
         "ress": "^5.0.2",
-        "vue": "^2.6.14",
-        "vue-banana-i18n": "1.5.0",
-        "vuex": "^3.6.2"
-    },
-    "overrides": {
-        "@wikimedia/codex": {
-            "vue": "^2.6.14"
-        }
+        "vue": "^3.3.8",
+        "vue-banana-i18n": "2.3.0"
     }
 }
