@@ -1,90 +1,74 @@
 <template>
-	<div class="languageSelector__input__wrapper">
-		<div class="languageSelector__input-left-side">
-			<div class="languageSelector__input__search-icon">
-				<img :src="searchUrl" alt="">
-			</div>
-			<input
-				ref="input"
-				type="text"
-				class="languageSelector__input"
-				:value="value"
-				:placeholder="placeholder"
-				@input="onInput"
-				@keydown.tab="onTab"
-				@keydown.down.prevent="onArrowDown"
-				@keydown.up.prevent="onArrowUp"
-				@keydown.enter="onEnter"
-				@keydown.esc.prevent="onEscape"
-			>
-		</div>
-		<button
-			class="languageSelector__input__clear-button"
-			:class="clearBtnVisible ? 'languageSelector__input__clear-button--visible' : ''"
-			@click="onClearInputValue"
-		>
-			<img :src="clearUrl" :alt="$i18n( 'language-selector-clear-button-label' )">
-		</button>
-	</div>
+  <div class="languageSelector__input__wrapper">
+    <div class="languageSelector__input-left-side">
+      <div class="languageSelector__input__search-icon">
+        <img
+          :src="searchUrl"
+          alt=""
+        >
+      </div>
+      <input
+        ref="input"
+        type="text"
+        class="languageSelector__input"
+        :value="value"
+        :placeholder="placeholder"
+        @input="onInput"
+        @keydown.tab="$emit('tab')"
+        @keydown.down.prevent="$emit('arrowDown')"
+        @keydown.up.prevent="$emit('arrowUp')"
+        @keydown.enter="$emit('enter')"
+        @keydown.esc.prevent="$emit('escape')"
+      >
+    </div>
+    <button
+      class="languageSelector__input__clear-button"
+      :class="clearBtnVisible ? 'languageSelector__input__clear-button--visible' : ''"
+      @click="onClearInputValue"
+    >
+      <img
+        :src="clearUrl"
+        :alt="$i18n( 'language-selector-clear-button-label' )"
+      >
+    </button>
+  </div>
 </template>
 
-<script lang="ts">
-import Vue from 'vue';
-import searchUrl from '../../img/search.svg';
-import clearUrl from '../../img/clear.svg';
+<script setup lang="ts">
+import {ref, computed} from 'vue';
+import searchUrlSvg from '../../img/search.svg';
+import clearUrlSvg from '../../img/clear.svg';
 
-export default Vue.extend( {
-	name: 'LanguageSelectorInput',
-	props: {
-		value: {
-			type: String,
-			default: '',
-		},
-		placeholder: {
-			type: String,
-			default: '',
-		},
-	},
-	data() {
-		return {
-			searchUrl,
-			clearUrl,
-		};
-	},
-	computed: {
-		clearBtnVisible(): boolean {
-			return this.value.length > 0;
-		},
-	},
-	methods: {
-		onClearInputValue(): void {
-			this.$emit( 'clear' );
-			this.focus();
-		},
+const props = defineProps<{
+	value: string,
+	placeholder: string
+}>();
 
-		focus(): void {
-			( this.$refs.input as HTMLInputElement ).focus();
-		},
-		onArrowDown() {
-			this.$emit( 'arrowDown' );
-		},
-		onArrowUp() {
-			this.$emit( 'arrowUp' );
-		},
-		onEnter() {
-			this.$emit( 'enter' );
-		},
-		onEscape() {
-			this.$emit( 'escape' );
-		},
-		onInput( event : Event ) {
-			this.$emit( 'input', (event.target as HTMLInputElement).value );
-		},
-		onTab() {
-			this.$emit( 'tab' );
-		},
-	},
-} );
+const emit = defineEmits(['clear', 'arrowDown', 'arrowUp', 'enter', 'escape', 'input', 'tab'])
+
+const searchUrl = ref(searchUrlSvg);
+const clearUrl = ref(clearUrlSvg);
+
+const input = ref<HTMLInputElement | null>(null);
+
+const clearBtnVisible = computed<boolean>(() => {
+	return props.value.length > 0;
+})
+
+function focus(): void {
+	(input.value as HTMLInputElement).focus();
+}
+
+function onClearInputValue(): void {
+	emit('clear');
+	focus();
+}
+
+function onInput(event: Event) {
+	emit('input', (event.target as HTMLInputElement).value);
+}
+
+defineExpose({focus});
 </script>
 
 <style lang="scss">
@@ -95,7 +79,7 @@ export default Vue.extend( {
 	font-weight: 400;
 	box-sizing: border-box;
 	flex-grow: 1;
-	border-color: #3366cc;
+	border-color: #36c;
 	height: 20px;
 
 	&:focus {
@@ -107,24 +91,23 @@ export default Vue.extend( {
 	}
 
 	&__wrapper {
-		background-color: #ffffff;
-		border-color: #a2a9b1;
+		background-color: #fff;
 		border-style: solid;
 		border-width: 1px;
-		border-radius: 2px 2px 0px 0px;
+		border-radius: 2px 2px 0 0;
 		padding-inline: 16px;
 		padding-block: 16px;
 		width: 100%;
 		display: flex;
 		justify-content: space-between;
 		box-shadow: 0 1px 2px #00000040, inset 0 0 0 1px #36c;
-		border-color: #3366cc;
+		border-color: #36c;
 		align-items: center;
 	}
 
 	&-left-side {
-		display:flex;
-		flex-grow:1;
+		display: flex;
+		flex-grow: 1;
 	}
 
 	&__search-icon {
@@ -134,7 +117,6 @@ export default Vue.extend( {
 	}
 
 	&__clear-button {
-
 		visibility: hidden;
 		display: flex;
 
