@@ -6,12 +6,12 @@ module.exports = {
 	extends: [
 		'eslint:recommended',
 		'@vue/typescript/recommended',
-		'plugin:vue/essential',
+		'plugin:vue/vue3-strongly-recommended',
 	],
 	parser: 'vue-eslint-parser',
 	rules: {
 		'max-len': [ 'error', 120 ],
 		'no-multiple-empty-lines' : ['error', { 'max': 1 }],
 		'vue/multi-word-component-names' : [ 'off' ]
-		},	
+		},
 };
