@@ -46,15 +46,57 @@ class ImportCSV implements ShouldQueue
         $filepath = Storage::disk('local')
             ->path('mismatch-files/' . $this->meta->filename);
 
-        DB::transaction(function () use ($reader, $filepath) {
-            $reader->lines($filepath)->each(function ($mismatchLine) {
-                $mismatch = Mismatch::make($mismatchLine);
-                if ($mismatch->type == null) {
-                    $mismatch->type = 'statement';
+        $mismatch_attrs = (new Mismatch())->getFillable();
+
+        DB::transaction(function () use ($reader, $filepath, $mismatch_attrs) {
+            $new_mismatches = [];
+            $where_clauses = [];
+
+            $mismatches_per_upload_user = DB::table('mismatches')
+                ->select($mismatch_attrs)
+                ->join('import_meta', 'mismatches.import_id', '=', 'import_meta.id')
+                ->where('import_meta.user_id', '=', $this->meta->user->id);
+
+            $reader->lines($filepath)->each(function ($mismatchLine) use (
+                &$new_mismatches,
+                &$where_clauses
+            ) {
+
+                $new_mismatch = $this->createMismatch($mismatchLine);
+                $new_mismatches[] = $new_mismatch;
+                $where_clause = [['review_status', '!=', 'pending']];
+                foreach ($new_mismatch->getAttributes() as $key => $attribute) {
+                    if ($key == 'review_status') {
+                        continue;
+                    }
+                    $where_clause[] = [$key, $attribute];
                 }
-                $mismatch->importMeta()->associate($this->meta);
-                $mismatch->save();
+                $where_clauses[] = $where_clause;
             });
+
+            $mismatches_per_upload_user->where(function ($query) use ($where_clauses) {
+                foreach ($where_clauses as $where_clause) {
+                    $query->orWhere(function ($query) use ($where_clause) {
+                        $query->where($where_clause);
+                    });
+                }
+            });
+
+            $existing_mismatches = $mismatches_per_upload_user->get();
+
+            foreach ($new_mismatches as $new_mismatch) {
+                if ($existing_mismatches->doesntContain(function ($value) use ($new_mismatch) {
+                    $metaAttrs = $new_mismatch->getAttributes();
+                    foreach ($metaAttrs as $attrKey => $attr) {
+                        if ($attrKey != 'review_status' && $value->{$attrKey} != $attr) {
+                            return false;
+                        }
+                    }
+                    return true;
+                })) {
+                    $this->saveMismatch($new_mismatch);
+                }
+            }
 
             $this->meta->status = 'completed';
             $this->meta->save();
@@ -64,7 +106,7 @@ class ImportCSV implements ShouldQueue
     /**
      * Handle a job failure.
      *
-     * @param  \Throwable  $exception
+     * @param \Throwable $exception
      * @return void
      */
     public function failed(Throwable $exception)
@@ -77,5 +119,30 @@ class ImportCSV implements ShouldQueue
 
         $this->meta->status = 'failed';
         $this->meta->save();
+    }
+
+
+    private function createMismatch($mismatch_data)
+    {
+        $new_mismatch = Mismatch::make($mismatch_data);
+
+        if ($new_mismatch->type == null) {
+            $new_mismatch->type = 'statement';
+        }
+
+        return $new_mismatch;
+    }
+
+    /**
+     * Save mismatch to database
+     *
+     * @param \Mismatch $new_mismatch
+     * @return void
+     */
+    private function saveMismatch($new_mismatch)
+    {
+        // if review_status == pending -> save
+        $new_mismatch->importMeta()->associate($this->meta);
+        $new_mismatch->save();
     }
 }
