@@ -1,47 +1,59 @@
 <template>
-    <div class="auth-widget" v-if="user">
-        <WikitLink :href="`https://www.wikidata.org/wiki/User:${user.name}`">
-            <img src="images/user.svg" class="icon-user" /><span class="username">{{ user.name }}</span>
-        </WikitLink>
-        <WikitLink href="/auth/logout">{{ $i18n('log-out') }}</WikitLink>
+  <div class="auth-widget_wrapper"> 
+    <div
+      class="auth-widget"
+      v-if="user"
+    >
+      <a :href="`https://www.wikidata.org/wiki/User:${user.name}`">
+        <img
+          src="images/user.svg"
+          class="icon-user"
+        ><span class="username">{{ user.name }}</span>
+      </a>
+      <a href="/auth/logout">{{ $i18n('log-out') }}</a>
     </div>
-    <WikitLink v-else href="/auth/login">{{ $i18n('log-in') }}</WikitLink>
+    <a
+      v-else
+      href="/auth/login"
+    >{{ $i18n('log-in') }}</a>
+  </div>
 </template>
 
-<script lang="ts">
+<script setup lang="ts">
 
-import Vue, { PropType } from 'vue';
 import User from '../types/User';
-import { Link as WikitLink } from '@wmde/wikit-vue-components';
 
-export default Vue.extend({
-    props: {
-        user: Object as PropType<User>
-    },
-    components: {
-        WikitLink
-    }
-});
+defineProps<{ user: User | null}>();
+
 </script>
 
 <style lang="scss">
-@import '~@wmde/wikit-tokens/dist/_variables.scss';
+@import '~@wikimedia/codex-design-tokens/theme-wikimedia-ui';
+@import '../../css/custom-variables.css';
 
 .auth-widget {
     display: flex;
-    color: $color-base-50;
-
-    .username {
-        margin-inline-end: $dimension-spacing-xlarge;
-    }
+    justify-content: space-between;
+    gap: var(--dimension-spacing-xlarge);
+    align-items: center;
 
     .icon-user {
-        margin-inline-end: $dimension-spacing-small;
+        margin-inline-end: var(--dimension-spacing-small);
         vertical-align: middle;
     }
 
-    .wikit-Link {
-        white-space: nowrap;
+    &_wrapper {
+        a {
+            white-space: nowrap;
+
+            &:visited {
+                color: $color-progressive;
+
+                &:hover {
+                    color: $color-progressive--hover;
+                }
+            }
+        }
     }
 }
 
