@@ -1,125 +1,96 @@
 <template>
-    <div class="loading-indicator" v-if="shown">
-        <div class="progressbar" role="progressbar" />
-        <div class="overlay" />
+  <div
+    class="loading-indicator"
+    v-if="shown"
+  >
+    <div class="overlay">
+      <cdx-progress-bar aria-label="Indeterminate progress bar" />
     </div>
+  </div>
 </template>
 
-<script lang="ts">
-import Vue from 'vue';
+<script setup lang="ts">
+import { ref } from 'vue';
+import { CdxProgressBar } from "@wikimedia/codex";
 
-export default Vue.extend({
-    name: 'LoadingOverlay',
-    props: {
-        delay: {
-            type: Number,
-            default: 250,
-        },
-        visible: {
-            type: Boolean,
-            default: false,
-        }
-    },
-    data() {
-        return {
-            shown: this.visible,
-            cachedStyles: {
-                overflow: 'auto'
-            }
-        };
-    },
-    methods: {
-        show(): void {
-            if (this.shown) {
-                return;
-            }
-
-            // Determine the current styles for body, in order to cache the overflow
-            const bodyStyles = window.getComputedStyle(document.body);
-            this.shown = true;
-
-            this.cachedStyles.overflow = bodyStyles.overflow;
-            document.body.style.overflow = 'hidden';
-        },
-        hide(): Promise<void> {
-            return new Promise((resolve) => {
-                setTimeout(() => {
-                    this.shown = false;
-                    // Restore previous overflow value
-                    document.body.style.overflow = this.cachedStyles.overflow;
-                    resolve();
-                }, this.delay);
-            });
-        },
-    },
+const props = withDefaults(defineProps<{
+	delay?: number
+	visible?: boolean
+}>(), {
+	delay: 250,
+	visible: false
 });
+
+const shown = ref(props.visible);
+const cachedStyles = ref({ overflow: 'auto' });
+
+function show(){
+    if (shown.value) {
+        return;
+    }
+
+    // Determine the current styles for body, in order to cache the overflow
+    const bodyStyles = window.getComputedStyle(document.body);
+    shown.value = true;
+
+    cachedStyles.value.overflow = bodyStyles.overflow;
+    document.body.style.overflow = 'hidden';
+}
+
+function hide(): Promise<void> {
+    return new Promise((resolve) => {
+        setTimeout(() => {
+            shown.value = false;
+            // Restore previous overflow value
+            document.body.style.overflow = cachedStyles.value.overflow;
+            resolve();
+        }, props.delay);
+    });
+}
+
+defineExpose({show, hide});
 </script>
 
 <style lang="scss">
-@import '~@wmde/wikit-tokens/dist/_variables.scss';
+@import "@wikimedia/codex-design-tokens/theme-wikimedia-ui";
 
 $base: '.loading-indicator';
 
-#{$base} .overlay {
-    /**
-    * Layout
-    */
-    width: $wikit-Dialog-overlay-width;
-    height: $wikit-Dialog-overlay-height;
-    position: fixed;
-    top:0;
-    left:0;
-
-    z-index: 100;
-
-    /**
-    * Colors
-    */
-    background-color: $wikit-Dialog-overlay-background-color;
-    opacity: $wikit-Dialog-overlay-opacity;
-}
-
-#{$base} .progressbar {
-    // Currently the inline progress bar only supports indeterminate loading mode.
-    // For a proof of concept on how this can include also determinate loading, see:
-    // https://codepen.io/xumium/pen/LYLZbva?editors=1100
-    // We ensure semantic usage by only targeting generic elements that set the
-    // correct role
-    &[role=progressbar] {
+#{$base} {
+    .overlay {
+        /**
+        * Layout
+        */
+        display: flex;
+        align-items: center;
+        justify-content: center;
         position: fixed;
         top: 0;
         left: 0;
-        width: $wikit-Progress-inline-track-width;
-        height: $wikit-Progress-inline-track-height;
+        z-index: $z-index-overlay-backdrop;
+        min-height: $size-full;
+        width: $size-viewport-width-full;
+        height: $size-viewport-height-full;
 
-        &::before {
-            content: '';
-            position: absolute;
-            top: 0;
-            left: 0;
-            display: block;
-            height: 100%;
-            background: $wikit-Progress-inline-background-color;
-        }
+        // Support Safari/iOS: Make `100vh` work with Safari's address bar.
+        // See https://stackoverflow.com/questions/37112218/css3-100vh-not-constant-in-mobile-browser
+        /* stylelint-disable-next-line plugin/no-unsupported-browser-features,
+            scale-unlimited/declaration-strict-value */
+        height: stretch;
 
-        // Indeterminate progress bars should not set the `aria-valuenow`
-        // attribute
-        &:not([aria-valuenow])::before {
-            width: 30%;
-            border-radius: $wikit-Progress-inline-indeterminate-border-radius;
-            animation-name: load-indeterminate;
-            animation-duration: $wikit-Progress-inline-animation-duration;
-            animation-timing-function: ease;
-            animation-iteration-count: infinite;
-            animation-delay: 0s;
-        }
+        /**
+        * Colors
+        */
+        background-color: $background-color-backdrop-light;
+
     }
 
-    @keyframes load-indeterminate {
-        0% { left: 0; }
-        50% { left: 70%; }
-        100% { left: 0; }
+    .cdx-progress-bar {
+        min-width: 432px;
+
+        @media (max-width: $max-width-breakpoint-tablet) {
+            min-width: 90vw;
+        }
     }
-    z-index: 101;
 }
 </style>
