@@ -1,13 +1,12 @@
 declare module "*.vue" {
-    import Vue from 'vue'
-
     module 'vue/types/vue' {
         interface Vue {
             $i18n: ( msg: string, ...args: unknown[] ) => string
         }
     }
-
-    export default Vue
+    import { defineComponent } from "vue";
+    const component: ReturnType<typeof defineComponent>;
+    export default component;
 }
 
 declare module 'vue-banana-i18n';
@@ -17,4 +16,4 @@ declare module '@wikimedia/language-data';
 declare module '*.svg' {
     const content: any;
     export default content;
-  }
+}
